requirements.txt
flask
flask_sqlalchemy
flask_migrate
python-dotenv
tweepy
textblob
deep-translator
pymysql
